Development - implementation of messageHandlers for each Activity
as of now messageHandlers I added just print the message to the log,
but the handlers should be adjusted to load the message in JSON then
get data from it for use by the Activity

Also debugged issue with set_player_name JSON payload not matching what server
expected, and confirmed successful receipt of response.

Added one line to TicTacToeActivity to remove the previous handler before
adding its handler.